Create sample_profile.json
This is not a complete profile file, but sample. The highlights are: it is hierarchical to reflect x.509 certificate key hierarchic and also importantly the keys should be the same as the keys in ASN1Crytpo library. The keys are shared between profile and cert. A complete profile will be generated soon.